app/requirements.txt
Flask
RPi.GPIO
pyRF24
pynmea2
